requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Diffusion Models
diffusers>=0.30.0
transformers>=4.38.0
accelerate>=0.20.0

# Model Loading and Storage
safetensors>=0.4.0
huggingface_hub>=0.20.0

# API Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
websockets>=11.0

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
numpy>=1.24.0
requests>=2.31.0
tqdm>=4.65.0
psutil>=5.9.0

# Optional Video Generation
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# Development Tools (Optional)
pytest>=7.4.0
black>=23.0.0